Fix broken links by omitting them
Unfortunately in the case of Profiles, some example relied on HiPE which
is now deprecated. I've replaced the early examples with a deterministic
compiler option, but HiPE was still require for the lower section using
fancier/more complex tuples.

I've instead denoted it as deprecated while still useful to show the
principles.